+++ R/ggseqplot-package.R
@@ -1,2 +1,3 @@
 #' @keywords internal
+#' @import TraMineR
 "_PACKAGE"